+++ extensions/lua/ext/fpconv.h
@@ -4,15 +4,17 @@
  *
  * Longest double printed with %.14g is 21 characters long:
  * -1.7976931348623e+308 */
+#include "pconf.h"
+
 # define FPCONV_G_FMT_BUFSIZE   32
 
 #ifdef USE_INTERNAL_FPCONV
-static inline void fpconv_init()
+STIN void fpconv_init()
 {
     /* Do nothing - not required */
 }
 #else
-extern inline void fpconv_init();
+EXIN void fpconv_init();
 #endif
 
 extern int fpconv_g_fmt(char*, double, int);