+++ extensions/lua/ext/strbuf.h
@@ -25,6 +25,8 @@
 #include <stdlib.h>
 #include <stdarg.h>
 
+#include "pconf.h"
+
 /* Size: Total bytes allocated to *buf
  * Length: String length, excluding optional NULL terminator.
  * Increment: Allocation increments when resizing the string buffer.
@@ -75,74 +77,74 @@ static void strbuf_append_char(strbuf_t *s, const char c);
 static void strbuf_ensure_null(strbuf_t *s);
 
 /* Reset string for before use */
-static inline void strbuf_reset(strbuf_t *s)
+STIN void strbuf_reset(strbuf_t *s)
 {
     s->length = 0;
 }
 
-static inline int strbuf_allocated(strbuf_t *s)
+STIN int strbuf_allocated(strbuf_t *s)
 {
     return s->buf != NULL;
 }
 
 /* Return bytes remaining in the string buffer
  * Ensure there is space for a NULL terminator. */
-static inline int strbuf_empty_length(strbuf_t *s)
+STIN int strbuf_empty_length(strbuf_t *s)
 {
     return s->size - s->length - 1;
 }
 
-static inline void strbuf_ensure_empty_length(strbuf_t *s, int len)
+STIN void strbuf_ensure_empty_length(strbuf_t *s, int len)
 {
     if (len > strbuf_empty_length(s))
         strbuf_resize(s, s->length + len);
 }
 
-static inline char *strbuf_empty_ptr(strbuf_t *s)
+STIN char *strbuf_empty_ptr(strbuf_t *s)
 {
     return s->buf + s->length;
 }
 
-static inline void strbuf_extend_length(strbuf_t *s, int len)
+STIN void strbuf_extend_length(strbuf_t *s, int len)
 {
     s->length += len;
 }
 
-static inline int strbuf_length(strbuf_t *s)
+STIN int strbuf_length(strbuf_t *s)
 {
     return s->length;
 }
 
-static inline void strbuf_append_char(strbuf_t *s, const char c)
+STIN void strbuf_append_char(strbuf_t *s, const char c)
 {
     strbuf_ensure_empty_length(s, 1);
     s->buf[s->length++] = c;
 }
 
-static inline void strbuf_append_char_unsafe(strbuf_t *s, const char c)
+STIN void strbuf_append_char_unsafe(strbuf_t *s, const char c)
 {
     s->buf[s->length++] = c;
 }
 
-static inline void strbuf_append_mem(strbuf_t *s, const char *c, int len)
+STIN void strbuf_append_mem(strbuf_t *s, const char *c, int len)
 {
     strbuf_ensure_empty_length(s, len);
     memcpy(s->buf + s->length, c, len);
     s->length += len;
 }
 
-static inline void strbuf_append_mem_unsafe(strbuf_t *s, const char *c, int len)
+STIN void strbuf_append_mem_unsafe(strbuf_t *s, const char *c, int len)
 {
     memcpy(s->buf + s->length, c, len);
     s->length += len;
 }
 
-static inline void strbuf_ensure_null(strbuf_t *s)
+STIN void strbuf_ensure_null(strbuf_t *s)
 {
     s->buf[s->length] = 0;
 }
 
-static inline char *strbuf_string(strbuf_t *s, int *len)
+STIN char *strbuf_string(strbuf_t *s, int *len)
 {
     if (len)
         *len = s->length;