+++ pulsar/utils/__init__.py
@@ -76,4 +76,5 @@ Path
 
 .. automodule:: pulsar.utils.path
    :members:
+
 '''